fix: detection of challenge/response
previously we thought the protocol version is enough to identify
if the KILO etc is needed or not. Unfortunately this is not true as
we have devices like the H811 running on MM (20v) which requires
KILO but still using protocol 1.0.0.1 !
Same goes for some H815 firmwares running MM (20g) while other
MM versions on H815 (20p) have the newer protocol implemented.
regardless of which firmware flashed the need of using KILO or not
can be done on the G4 by the USB product ID (633e do not need it
while 633a need it).

Adding other USB product IDs who needs KILO can be simply done
by adding them to the new function:
kilo_lg_product_ids