identify block size automatically (UFS/EMMC)
based on the assumption that there are just  a small amount of
possible block sizes (can be easily extended though) and that we
can find the EFI PART signature there.
Based on where we found the GPT signature the corresponding
block size and lba length will be used.

This also adds the possibility to ENFORCE a device type (and so the
block size) if the detection ever fails by using
--devtype [UFS|EMMC]

The debug output has been extended to print all identified
stuff and vars used for the current connection as well