new: added --proto [PROTO]
this will allow to specify a protocol version on the CLI.
usually lglaf auto negotiates the minimal allowed version reported
by the device but when specifying --proto [PROTO] that will be
ignored and the given version is used instead.

Format:
--proto 0x1000003 for version 3
--proto 0x1000018 for version 18
etc.

The format must be exactly like that and will be checked